requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djoser>=2.2
mysqlclient>=2.2
